+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0 (unreleased)
 
 - Added `Stl` struct
+- Added `into_parts` function
 - Changed `StlResult` methods to return slices
 
 ## 0.2.2 (2023-06-20)